Disband tight interaction between the task and WebRTC data channels
Data channel instances are no longer handled by the task. The
application has to take care of managing data channels itself. If the
application wants to apply the established SaltyRTC session keys, it can
use a DataChannelCryptoContext instance to encrypt and decrypt data.

Furthermore, if the application wants to hand over the signalling
channel to a dedicated data channel, that data channel must also be
managed by the application itself while the SignalingTransport will take
care of the crypto and the chunking.

In addition, the maximum packet size negotiation has been removed.