Refactor signalling transport handover API
The previous proposal was insufficient since messages could be lost
between the underlying data channel opening and resolving the transport
factory promise. In addition, it was impossible to trigger the handover
process.

Furthermore, the SignalingTransportHandler has been split into two
entities where the SignalingTransportHandler is implemented by the
application and the SignalingTransportLink is provided by the task
towards the application. This makes it less awkward to work with.